Minor imports optimization

+++ src/main/java/io/github/persiancalendar/praytimes/Utils.java
@@ -14,7 +14,7 @@ class Utils {
     }
 
     public static MinuteOrAngleDouble min(int value) {
-        return new MinuteOrAngleDouble((double) value, true);
+        return new MinuteOrAngleDouble(value, true);
     }
 
     static MinuteOrAngleDouble deg(double value) {
